Week1 [STEP 3] Kobe
 @Gundy93
 건디 수정사안 모두 반영하여 다시 PR 보냅니다 :)

 ## 고민했던 점
 ### 분기처리
 - nil 이 들어옴에 따라 빼빼로의 모양이 달라지는데 그것에 따른 분기 처리를 어떻게 할지 고민해봤습니다.

 ---

 ## 해결이 되지 않은 점
 ### 이번에는 질문-답변 채널에서 답을 얻었습니다.
 - main.swift 파일에서 직접 STEP3.swift 함수를 호출해야 한다는 점을 알게 되었습니다.

 ---

 ## 조언을 얻고 싶은 부분
 ### 조바심이 나는 부분
 - 아무래도 현업과 병행을 하다보니 다른 캠퍼분들과 달리 진도가 매우 느린것 같아 조바심이 납니다.
 다른 캠퍼분들은 이미 3주차를 진행하는데 저는 1주차도 못끝낸것 같아 매우 조바심이 납니다. ㅠㅠ
